app/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pydantic
python-multipart
